+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ## 0.55 - TBD - Refinements
-* Made it possible to enter time entrie in the past (if time is given)
+* Made it possible to enter time entries in the past (if time is given)
+* Stats-bar for "types" are shown.
 
 ## 0.54 - 07.09.2017 - Refinements
 * Animation improvements