+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 0.55 - TBD - Refinements
 * Made it possible to enter time entries in the past (if time is given)
 * Stats-bar for "types" are shown.
+* More component driven setup
+* Started with Project-component
 
 ## 0.54 - 07.09.2017 - Refinements
 * Animation improvements